pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-converter"
version = "0.1.0"
description = "Local-only file converter with plugin-based conversion engine"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "File Converter Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "flet>=0.21.0",
    "python-magic>=0.4.27",
    "pluggy>=1.3.0",
    "tomli>=2.0.1",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
]

[project.scripts]
fc = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli*"]

[tool.setuptools.package-data]
file_converter = ["plugins/*/plugin.toml"]

